+++ src/main/scala/dxWDL/base/Utils.scala
@@ -34,6 +34,11 @@ object Utils {
     val MAX_NUM_FILES_MOVE_LIMIT = 1000
     val UBUNTU_VERSION = "16.04"
     val VERSION_PROP = "dxWDL_version"
+    val REORG_CONFIG = "___reorg_conf"
+    val REORG_STATUS = "___reorg_status"
+    val REORG_STATUS_COMPLETE = "completed"
+
+
     var traceLevel = 0
 
     // The version lives in application.conf
@@ -355,3 +360,10 @@ object Utils {
         "[" + concat + "]"
     }
 }
+
+
+class PermissionDeniedException(s:String) extends Exception(s){}
+
+class InvalidInputException(s:String) extends Exception(s){}
+
+class IllegalArgumentException(s:String) extends Exception(s){}