+++ src/main/scala/dxWDL/compiler/Native.scala
@@ -320,12 +320,16 @@ case class Native(dxWDLrtId: Option[String],
         val body:String = applet.kind match {
             case IR.AppletKindNative(_) =>
                 throw new Exception("Sanity: generating a bash script for a native applet")
+            case IR.AppletKindWorkflowCustomReorg(_) =>
+                throw new Exception("Sanity: generating a bash script for a custom reorg applet")
             case IR.AppletKindWfFragment(_, _, _) =>
                 genBashScriptWfFragment()
             case IR.AppletKindWfInputs =>
                 genBashScriptCmd("wfInputs")
             case IR.AppletKindWfOutputs =>
                 genBashScriptCmd("wfOutputs")
+            case IR.AppletKindWfCustomReorgOutputs =>
+                genBashScriptCmd("wfCustomReorgOutputs")
             case IR.AppletKindWorkflowOutputReorg =>
                 genBashScriptCmd("workflowOutputReorg")
             case IR.AppletKindTask(_) =>
@@ -667,6 +671,7 @@ case class Native(dxWDLrtId: Option[String],
 
                 case IR.AppletKindWfInputs |
                         IR.AppletKindWfOutputs |
+                        IR.AppletKindWfCustomReorgOutputs |
                         IR.AppletKindWorkflowOutputReorg =>
                     // meta information used for running workflow fragments
                     val fqnDictTypes = JsObject(
@@ -1033,6 +1038,9 @@ case class Native(dxWDLrtId: Option[String],
                         val execRecord = apl.kind match {
                             case IR.AppletKindNative(id) =>
                                 // native applets do not depend on other data-objects
+                                ExecRecord(apl, DxExec(id), Vector.empty)
+                            case IR.AppletKindWorkflowCustomReorg(id) =>
+                                // does this has to be a different class?
                                 ExecRecord(apl, DxExec(id), Vector.empty)
                             case _ =>
                                 val (dxApplet, dependencies) = buildAppletIfNeeded(apl, accu)