+++ src/main/scala/dxWDL/compiler/Top.scala
@@ -237,8 +237,19 @@ case class Top(cOpt: CompilerOptions) {
             case None => WdlRuntimeAttrs(Map.empty)
             case Some(ex) => ex.defaultRuntimeAttributes
         }
+
+        val reorgApp: Either[Boolean, ReorgAttrs] = cOpt.extras match {
+
+            case None => Left(cOpt.reorg)
+            case Some(ex) => ex.customReorgAttributes match {
+                case None => Left(cOpt.reorg)
+                case Some(cOrg) => Right(cOrg)
+            }
+
+        }
+
         new GenerateIR(cOpt.verbose, defaultRuntimeAttrs).apply(everythingBundle, allSources, language,
-                                                                cOpt.locked, cOpt.reorg)
+                                                                cOpt.locked, reorgApp)
     }
 
     // Compile IR only