+++ src/test/scala/dxWDL/base/ExtrasTest.scala
@@ -1,14 +1,22 @@
 package dxWDL.base
 
 import com.dnanexus.AccessLevel
+import com.dnanexus.exceptions.ResourceNotFoundException
 import dxWDL.compiler.EdgeTest
 import org.scalatest.{FlatSpec, Matchers}
 import spray.json._
 import wom.values._
+import DefaultJsonProtocol._
 
 class ExtrasTest extends FlatSpec with Matchers {
     val verbose = Verbose(true, true, Set.empty)
 
+    private def getIdFromName(name: String): String = {
+      val (stdout, stderr) = Utils.execCommand(s"dx describe ${name} --json")
+      stdout.parseJson.asJsObject match {
+        case JsObject(x) => JsObject(x).fields("id").convertTo[String]
+      }
+    }
 
     it should "recognize restartable entry points" in {
         val runtimeAttrs : JsValue =
@@ -185,10 +193,233 @@ class ExtrasTest extends FlatSpec with Matchers {
         }
     }
 
+    it should "parse the custom_reorg object" in {
 
-    it should "generate valid JSON execution policy" in {
-        val expectedJs : JsValue =
-            """|{
+        // app_id is mummer nucmer app in project-FJ90qPj0jy8zYvVV9yz3F5gv
+        val appId : String = getIdFromName("/release_test/mummer_nucmer_aligner")
+        val fileId : String = getIdFromName("Readme.md")
+
+        // inputs is Readme.md file in project-FJ90qPj0jy8zYvVV9yz3F5gv
+        val reorg: JsValue   =
+            s"""|{
+               | "custom_reorg" : {
+               |    "app_id" :"${appId}",
+               |    "conf" : "${fileId}"
+               |  }
+               |}
+               |""".stripMargin.parseJson
+
+        val extras = Extras.parse(reorg, verbose)
+        extras.customReorgAttributes  should be (
+            Some(ReorgAttrs(appId, fileId))
+        )
+    }
+
+    it should "throw IllegalArgumentException due to missing applet id" in {
+
+        val inputs: String = "dx://file-123456"
+        val reorg: JsValue   =
+            s"""|{
+                | "custom_reorg" : {
+                |    "conf" : "${inputs}"
+                |  }
+                |}
+                |""".stripMargin.parseJson
+
+
+        val thrown = intercept[dxWDL.base.IllegalArgumentException] {
+            Extras.parse(reorg, verbose)
+        }
+
+        thrown.getMessage should be ("app_id must be specified in the custom_reorg section.")
+    }
+
+    it should "throw IllegalArgumentException due to missing inputs in custom_reorg section" in {
+
+        // app_id is mummer nucmer app in project-FJ90qPj0jy8zYvVV9yz3F5gv
+        val appId : String = getIdFromName("/release_test/mummer_nucmer_aligner")
+        val reorg: JsValue   =
+        s"""|{
+            | "custom_reorg" : {
+            |    "app_id" : "${appId}"
+            |  }
+            |}
+            |""".stripMargin.parseJson
+
+
+      val thrown = intercept[dxWDL.base.IllegalArgumentException] {
+          Extras.parse(reorg, verbose)
+      }
+
+      //thrown.getMessage should contain  ("inputs must be specified in the custom_reorg section.")
+      thrown.getMessage should be  (
+          "conf must be specified in the custom_reorg section. Please set the value to null if there is no conf file."
+      )
+    }
+
+    it should "Allow inputs to be null in custom reorg" in {
+
+      // app_id is mummer nucmer app in project-FJ90qPj0jy8zYvVV9yz3F5gv
+      val appId : String = getIdFromName("/release_test/mummer_nucmer_aligner")
+      val reorg: JsValue   =
+        s"""|{
+            | "custom_reorg" : {
+            |    "app_id" : "${appId}",
+            |    "conf" : null
+            |  }
+            |}
+            |""".stripMargin.parseJson
+
+
+      val extras = Extras.parse(reorg, verbose)
+      extras.customReorgAttributes should be (
+        Some(ReorgAttrs(appId, ""))
+      )
+  }
+
+  it should "throw IllegalArgumentException due to invalid applet ID" in {
+
+    // invalid applet ID
+    val appId : String = "applet-123456"
+    val reorg : JsValue =
+      s"""|{
+          |  "custom_reorg" : {
+          |      "app_id" : "${appId}",
+          |      "conf": null
+          |   }
+          |}
+          |""".stripMargin.parseJson
+
+    val thrown = intercept[dxWDL.base.IllegalArgumentException] {
+      Extras.parse(reorg, verbose)
+    }
+
+    thrown.getMessage should be  (
+      s"dxId must match applet-[A-Za-z0-9]{24}"
+    )
+
+  }
+
+  it should "throw ResourceNotFoundException due to non-existent applet" in {
+
+    // non-existent (made up) applet ID
+    val appId : String = "applet-mPX7K2j0Gv2K2jXF75Bf21v2"
+    val reorg : JsValue =
+      s"""|{
+          |  "custom_reorg" : {
+          |      "app_id" : "${appId}",
+          |      "conf": null
+          |   }
+          |}
+          |""".stripMargin.parseJson
+
+    val thrown = intercept[Exception] {
+      Extras.parse(reorg, verbose)
+    }
+
+    thrown.getMessage should be  (
+      s"""Error running command dx describe $appId --json"""
+    )
+
+  }
+
+  it should "throw IllegalArgumentException due to invalid file ID" in {
+
+    // app_id is mummer nucmer app in project-FJ90qPj0jy8zYvVV9yz3F5gv
+
+    val appId : String = getIdFromName("/release_test/mummer_nucmer_aligner")
+    val inputs : String = "file-1223445"
+    val reorg : JsValue =
+      s"""|{
+          |  "custom_reorg" : {
+          |      "app_id" : "${appId}",
+          |      "conf": "${inputs}"
+          |   }
+          |}
+          |""".stripMargin.parseJson
+
+    val thrown = intercept[java.lang.IllegalArgumentException] {
+      Extras.parse(reorg, verbose)
+    }
+
+    thrown.getMessage should be  (
+      s"dxId must match file-[A-Za-z0-9]{24}"
+    )
+
+  }
+
+  it should "throw ResourceNotFoundException due to non-existant file" in {
+
+    // app_id is mummer nucmer app in project-FJ90qPj0jy8zYvVV9yz3F5gv
+
+    val appId : String = getIdFromName("/release_test/mummer_nucmer_aligner")
+    val inputs : String = "dx://file-AZBYlBQ0jy1qpqJz17gpXFf8"
+    val reorg : JsValue =
+      s"""|{
+          |  "custom_reorg" : {
+          |      "app_id" : "${appId}",
+          |      "conf": "${inputs}"
+          |   }
+          |}
+          |""".stripMargin.parseJson
+
+    val thrown = intercept[ResourceNotFoundException] {
+      Extras.parse(reorg, verbose)
+    }
+
+    val fileId : String = inputs.replace("dx://", "")
+
+    thrown.getMessage should be  (
+      s""""${fileId}" is not a recognized ID"""
+    )
+
+  }
+
+  it should "throw PermissionDeniedException due to applet not having contribute access in the project" in {
+
+    // app_id is sum app in project-FJ90qPj0jy8zYvVV9yz3F5gv
+    val appId : String = getIdFromName("/release_test/Sum ")
+    val reorg: JsValue =
+      s"""|{
+          | "custom_reorg" : {
+          |    "app_id" : "${appId}",
+          |    "conf": "null"
+          |  }
+          |}
+          |""".stripMargin.parseJson
+
+    val thrown = intercept[PermissionDeniedException] {
+      Extras.parse(reorg, verbose)
+
+    }
+
+    thrown.getMessage should be (
+      s"ERROR: App(let) for custom reorg stage ${appId} does not " +
+        s"have CONTRIBUTOR or ADMINISTRATOR access and this is required."
+    )
+  }
+
+  it should "take app id as well as applet id for custom reorg" taggedAs (EdgeTest) in {
+
+    val appId : String = getIdFromName("cloud_workstation")
+    val reorg: JsValue =
+      s"""|{
+          | "custom_reorg" : {
+          |    "app_id" : "${appId}",
+          |    "conf": null
+          |  }
+          |}
+          |""".stripMargin.parseJson
+
+    val extras = Extras.parse(reorg, verbose)
+    extras.customReorgAttributes  should be (
+      Some(ReorgAttrs(appId, ""))
+    )
+  }
+
+  it should "generate valid JSON execution policy" in {
+    val expectedJs : JsValue =
+        """|{
                | "executionPolicy": {
                |    "restartOn": {
                |       "*": 5
@@ -536,7 +767,7 @@ class ExtrasTest extends FlatSpec with Matchers {
         result("upstreamProjects") should be (dxDetailsJson)
     }
 
-    it should "all DxAttr to return RunSpec Json" taggedAs(EdgeTest) in {
+    it should "all DxAttr to return RunSpec Json" in {
 
         val expectedPolicy = """
             |{
@@ -559,7 +790,7 @@ class ExtrasTest extends FlatSpec with Matchers {
 
     }
 
-    it should "all DxAttr to return empty runSpec and details Json" taggedAs(EdgeTest) in {
+    it should "all DxAttr to return empty runSpec and details Json" in {
 
         val dxAttrs = DxAttrs(None, None)