+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 1.36 18-Nov-2019
+
+Added a mechanism for a custom reorganization applet, it can be used instead of the built in --reorg option.
+You can use it to reorganize workflow file results after it completes.
+
+Please refer to [docs/ExpertOptions.md](docs/ExpertOptions.md#Adding-config-file-based-reorg-apple-at-compilation-time]
+
 ## 1.35.1
 - Object form of streaming syntax. This allows several annotations for an input/output
 parameter. In addition to the previously supported: